requirements.txt
# Core app
streamlit==1.37.0

# Data handling
pandas==2.2.2
numpy==1.26.4

# Document parsing
PyPDF2==3.0.1
python-pptx==0.6.23
openpyxl==3.1.5

# NLP & Semantic Search
sentence-transformers==2.2.2
faiss-cpu==1.8.0.post1

# Google Drive API
google-api-python-client==2.108.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
oauth2client==4.1.3

# Others
protobuf==4.25.3
requests==2.31.0

